Splitting out and optimising List.prototype.relationship
