+++ lib/list.js
@@ -115,6 +115,7 @@ function List(key, options) {
 }
 
 // Add prototype methods
+List.prototype.relationship = require('./list/relationship');
 List.prototype.getData = require('./list/getData');
 List.prototype.getOptions = require('./list/getOptions');
 List.prototype.addFiltersToQuery = require('./list/addFiltersToQuery');
@@ -339,40 +340,6 @@ Object.defineProperty(List.prototype, 'defaultColumns', {
 		delete this._defaultColumns;
 	}
 });
-
-/**
- * Registers relationships to this list defined on others
- */
-List.prototype.relationship = function(def) {
-	if (arguments.length > 1) {
-		_.each(arguments, function(def) { this.relationship(def); }, this);
-		return this;
-	}
-	if ('string' === typeof def) {
-		def = { ref: def };
-	}
-	if (!def.ref) {
-		throw new Error('List Relationships must be specified with an object containing ref (' + this.key + ')');
-	}
-	if (!def.refPath) {
-		def.refPath = utils.downcase(this.key);
-	}
-	if (!def.path) {
-		def.path = utils.keyToProperty(def.ref, true);
-	}
-	Object.defineProperty(def, 'refList', {
-		get: function() {
-			return keystone.list(def.ref);
-		}
-	});
-	Object.defineProperty(def, 'isValid', {
-		get: function() {
-			return keystone.list(def.ref) ? true : false;
-		}
-	});
-	this.relationships[def.path] = def;
-	return this;
-};
 
 
 /**
